pyproject.toml
[project]
name = "hrms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "dishka>=1.4.2",
    "fastapi[standard]>=0.115.8",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.5",
    "pydantic-settings>=2.7.1",
    "sqlalchemy>=2.0.37",
]
